backend/requirements.txt
# Core dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0

# Database
motor>=3.3.0
beanie>=1.25.0

# Vector DB & RAG
chromadb>=0.4.22
sentence-transformers>=2.3.0

# Agent framework
crewai>=0.80.0
google-genai
google-generativeai
# langchain dependencies are managed by crewai, but if needed explicitly:
langchain>=0.3.0
langchain-community>=0.3.0

# Background tasks
celery>=5.3.0
redis>=5.0.0

# Auth
firebase-admin>=6.4.0

# Utils
opik
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.26.0
